factory_tool/requirements.txt
bitarray==2.6.0
cryptography==36.0.2
cffi==1.15.0
future==0.18.2
pycparser==2.21
pypng==0.0.21
python_stdnum==1.18
ecdsa>=0.18.0
qrcode==7.4.2
pycryptodome>=3.17